+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed: `function-url-quotes` ignores values containing `$sass` and `@less` variables.
 - Fixed: selector-targeting rules ignore Less mixins and extends.
 - Fixed: `selector-type-no-unknown` ignores non-standard usage of percentage keyframe selectors (e.g. within an SCSS mixin).
+- Fixed: `no-extra-semicolons` reports errors on the correct line.
 
 # 6.5.1